Moved Exception based dtos into seperate dto package
